backend/requirements.txt
  #여기서 작업하고 커밋하고 push 한다음 바로 연준님께 보고 중간중간 조그마하게 푸쉬도 3,4개 쌓이면 바로푸쉬,백엔드디렉토리만 수정
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv